Keep only one typealias for SavedStateMap and moved it to utils-multiplatform
