+++ appyx-components/experimental/modal/common/src/commonMain/kotlin/com/bumble/appyx/components/modal/ModalModel.kt
@@ -4,9 +4,9 @@ import com.bumble.appyx.interactions.core.Element
 import com.bumble.appyx.interactions.core.Elements
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseTransitionModel
-import com.bumble.appyx.interactions.core.state.SavedStateMap
 import com.bumble.appyx.utils.multiplatform.Parcelable
 import com.bumble.appyx.utils.multiplatform.Parcelize
+import com.bumble.appyx.utils.multiplatform.SavedStateMap
 
 class ModalModel<InteractionTarget : Any>(
     initialElements: List<InteractionTarget>,
