+++ appyx-components/experimental/puzzle15/common/src/commonMain/kotlin/com/bumble/appyx/components/experimental/puzzle15/Puzzle15Model.kt
@@ -4,10 +4,10 @@ import com.bumble.appyx.components.experimental.puzzle15.Puzzle15Model.Tile
 import com.bumble.appyx.interactions.core.Element
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseTransitionModel
-import com.bumble.appyx.interactions.core.state.SavedStateMap
 import com.bumble.appyx.utils.multiplatform.Parcelable
 import com.bumble.appyx.utils.multiplatform.Parcelize
 import com.bumble.appyx.utils.multiplatform.RawValue
+import com.bumble.appyx.utils.multiplatform.SavedStateMap
 
 class Puzzle15Model(
     savedStateMap: SavedStateMap?
