+++ appyx-components/stable/backstack/common/src/commonMain/kotlin/com/bumble/appyx/components/backstack/BackStackModel.kt
@@ -5,9 +5,9 @@ import com.bumble.appyx.interactions.core.Element
 import com.bumble.appyx.interactions.core.Elements
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseTransitionModel
-import com.bumble.appyx.interactions.core.state.SavedStateMap
 import com.bumble.appyx.utils.multiplatform.Parcelable
 import com.bumble.appyx.utils.multiplatform.Parcelize
+import com.bumble.appyx.utils.multiplatform.SavedStateMap
 
 class BackStackModel<InteractionTarget : Any>(
     initialTargets: List<InteractionTarget>,
