--- appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/state/SavedStateMap.kt
@@ -1,6 +1,0 @@
-package com.bumble.appyx.interactions.core.state
-
-/**
- * Restored from [android.os.Bundle] state.
- */
-typealias SavedStateMap = Map<String, Any?>
