+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/permanent/PermanentModel.kt
@@ -4,10 +4,10 @@ import com.bumble.appyx.interactions.core.Element
 import com.bumble.appyx.interactions.core.Elements
 import com.bumble.appyx.interactions.core.asElement
 import com.bumble.appyx.interactions.core.model.transition.BaseTransitionModel
-import com.bumble.appyx.interactions.core.state.SavedStateMap
 import com.bumble.appyx.interactions.permanent.PermanentModel.State
 import com.bumble.appyx.utils.multiplatform.Parcelable
 import com.bumble.appyx.utils.multiplatform.Parcelize
+import com.bumble.appyx.utils.multiplatform.SavedStateMap
 
 class PermanentModel<InteractionTarget : Any>(
     savedStateMap: SavedStateMap?,
