+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/Node.kt
@@ -25,7 +25,7 @@ import com.bumble.appyx.navigation.plugin.NodeLifecycleAware
 import com.bumble.appyx.navigation.plugin.NodeReadyObserver
 import com.bumble.appyx.navigation.plugin.UpNavigationHandler
 import com.bumble.appyx.navigation.plugin.plugins
-import com.bumble.appyx.navigation.state.SavedStateMap
+import com.bumble.appyx.utils.multiplatform.SavedStateMap
 import com.bumble.appyx.navigation.store.RetainedInstanceStore
 import com.bumble.appyx.utils.multiplatform.BuildFlags
 import kotlinx.coroutines.CoroutineScope
